fix: validate function header type parsing

+++ tests/unit/test_il_comments.cpp
@@ -14,7 +14,7 @@ int main()
     const char *src = R"(// header line 1
 // header line 2
 il 0.1.2
-func @main() -> i32 {
+func @main() -> i64 {
 entry:
   ret 0
 }
